Remove some unused options from runTests.sh
The option -t for multiple core testing is currently not used.
This can be added at a later time if multiple core testing
should be done in one branch.

Alternative is branching out and only testing one TYPO3 version
in current branch.

Additionally, only composerInstall is used and composerInstallMax
will not be executed in the GitHub CI matrix.

Testing with sqllite is deactivated for now because it fails and
must be investigated further.