Fix bugs, TYPO3 v11 support, add functional tests and test setup (#9)
* Change version constraints to support TYPO3 v11

* Replace Command.php

Since TYPO3 v11, Services.yaml must be used to define
commands.

* Replace fetch() with appropriate function calls

For newer Doctrine dbal versions, the function calls
should be replaced:

- fetch() => fetchAssoicative()
- fetchAll() => fetchAllAssociative()
- fetchColumn(0) => fetchOne()

* [BUGFIX] Check for storage if checking for duplicates

Related: #2

* Add option --identifer to perform only on this identifier

Makes it more conveniant to check specific behaviour for
only a subset of files.

* Add option storage

* Add return 0 to execute()

Is necessary for TYPO3 v11

* Add additional output

* Make sure duplicate query works on case-insensitive DB collations

Since grouping by identifier and storage may falsely return
entries as duplicates where identifer is only the same if
compared case-insensitively, we group by identifier_hash as well.

Related: #2

* Update README

* Update comment

* Change duplicate check to do extra check in PHP

Du extra duplicate check in PHP instead of relying on identifier_hash.
This is a safer option, though it might be more inefficient.

Related: #2

* Perform a type-safe match

* Make sure nothing is performed if called with --dry-run

Simplify output messages.
Make sure nothing is done if command called with --dry-run.

* Return immediately if no duplicates found

If no duplicates found, the command exists with exit code 0.

If dupicates are found, it exits with exit code 1.

* Add testing-framework

* Add Test scripts

* Add GitHub CI script for running functional tests

* Remove TestEnvironment prepare

Is no longer necessary since TYPO3 v11.

* Add Resources/Public

This is only necessary for the test setup to work properly.

See https://review.typo3.org/c/Packages/TYPO3.CMS/+/71029

* Add functional tests

* Minor cleanup

* Remove rowCount()

This is not recommended in combination with SELECT
(see documentation) and it was not really necessary
as the while() statement is only entered if there
are results.

* Fix PHP version constraints

Use PHP versions supported by TYPO3 v11.

* Remove some unused options from runTests.sh

The option -t for multiple core testing is currently not used.
This can be added at a later time if multiple core testing
should be done in one branch.

Alternative is branching out and only testing one TYPO3 version
in current branch.

Additionally, only composerInstall is used and composerInstallMax
will not be executed in the GitHub CI matrix.

Testing with sqllite is deactivated for now because it fails and
must be investigated further.

* Add cleanBuild to runTests.sh

Cleanup can be performed with runTests.sh. The script
cleanup.sh is removed.

* Comment out some unused commands in runTest.sh

Configuration files for php-cs-fixer, unit test and phpstan are
not supplied yet. The commands are commented out in runTests.sh.